@@ -2,8 +2,8 @@ class Letter < ActiveRecord::Base
 
   belongs_to :user
   has_many :likes, dependent: :destroy
+  has_rich_text :description
   validates_presence_of :description
-
   def like_by?(user)
     likes.map(&:user_id).include?(user.id)
   end
