@@ -12,16 +12,19 @@ RSpec.describe Letter, type: :model do
   end
 
   describe 'methods in the class' do
-    let(:user) { create(:user) }
-    letter = Letter.create(description: 'Ruby is awesome')
+    before(:all) do
+      @user = create(:user)
+      @letter = create(:letter)
+    end
+
     context 'like_by?' do
       it 'should return false if user not liked' do
-        expect(letter.like_by?(user)).to eq(false)
+        expect(@letter.like_by?(@user)).to eq(false)
       end
 
       it 'should return true if user liked' do
-        letter.likes.create(user_id: user.id)
-        expect(letter.like_by?(user)).to eq(true)
+        @letter.likes.create(user_id: @user.id)
+        expect(@letter.like_by?(@user)).to eq(true)
       end
     end
   end
